# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Servicio que maneja la publicación y consulta de tweets.",
        "title": "tweets-service",
        "contact": {
            "name": "Fredy Hernandez",
            "email": "hfredy717@outlook.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8084",
    "basePath": "/",
    "paths": {
        "/tweets": {
            "post": {
                "description": "Permite a los usuarios publicar un nuevo tweet, almacenarlo en la base de datos y enviarlo a la cola de mensajes para su procesamiento asincrónico.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tweets"
                ],
                "summary": "Publicar un tweet",
                "parameters": [
                    {
                        "description": "Datos del tweet a publicar",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/kit.TweetRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Tweet creado correctamente",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Solicitud incorrecta",
                        "schema": {
                            "$ref": "#/definitions/kit.CommonApiError"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/kit.CommonApiError"
                        }
                    }
                }
            }
        },
        "/tweets/{user_id}/{tweet_id}": {
            "get": {
                "description": "Retorna un tweet basado en el `user_id` y `tweet_id` proporcionados.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tweets"
                ],
                "summary": "Obtener un tweet específico",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del usuario propietario del tweet",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID del tweet a consultar",
                        "name": "tweet_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tweet obtenido correctamente",
                        "schema": {
                            "$ref": "#/definitions/kit.Tweet"
                        }
                    },
                    "400": {
                        "description": "Solicitud incorrecta",
                        "schema": {
                            "$ref": "#/definitions/kit.CommonApiError"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/kit.CommonApiError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "kit.CommonApiError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Código del error.\nexample: \"ERR-400\"",
                    "type": "string"
                },
                "msg": {
                    "description": "Mensaje descriptivo del error.\nexample: \"Solicitud incorrecta\"",
                    "type": "string"
                }
            }
        },
        "kit.Tweet": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "Contenido del tweet.\nexample: \"Este es un tweet de prueba.\"",
                    "type": "string"
                },
                "created": {
                    "description": "Fecha de creación del tweet en formato UNIX timestamp.\nexample: 1710597823",
                    "type": "integer"
                },
                "tweet_id": {
                    "description": "ID único del tweet.\nexample: \"tweet:67890\"",
                    "type": "string"
                },
                "user_id": {
                    "description": "ID del usuario propietario del tweet.\nexample: \"user:12345\"",
                    "type": "string"
                }
            }
        },
        "kit.TweetRequest": {
            "type": "object",
            "required": [
                "tweet",
                "user_id"
            ],
            "properties": {
                "tweet": {
                    "description": "Contenido del tweet.\nRequired: true\nMax length: 280 caracteres\nexample: \"Este es un tweet de prueba con una longitud adecuada.\"",
                    "type": "string",
                    "maxLength": 280
                },
                "user_id": {
                    "description": "ID del usuario que publica el tweet.\nRequired: true\nexample: \"user:12345\"",
                    "type": "string"
                }
            }
        }
    }
}